+++ app/mapping/playback/UserAnswersExtractor.scala
@@ -19,6 +19,7 @@ package mapping.playback
 import com.google.inject.Inject
 import mapping.playback.PlaybackExtractionErrors._
 import mapping.playback.beneficiaries.BeneficiaryExtractor
+import mapping.playback.settlors.SettlorExtractor
 import models.playback.UserAnswers
 import models.playback.http.GetTrust
 import play.api.Logger
