+++ app/mapping/playback/settlors/SettlorCompanyExtractor.scala
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package mapping.playback
+package mapping.playback.settlors
 
 import com.google.inject.Inject
 import mapping.playback.PlaybackExtractionErrors.{FailedToExtractData, PlaybackExtractionError}
+import mapping.playback.{PlaybackExtractor, PlaybackImplicits}
 import models.core.pages.{IndividualOrBusiness, InternationalAddress, UKAddress}
 import models.playback.http.{DisplayTrustCharityType, DisplayTrustSettlorCompany}
 import models.playback.{MetaData, UserAnswers}
